+++ src/components/post/PostOlderThan.vue
@@ -12,6 +12,7 @@
 
 <script setup lang="ts">
 import type { localeStrings } from "@i18n/ui";
+
 import { useTranslations } from "@utils/i18n/utils";
 
 export interface PostOlderThanProps {