+++ src/tests/unit/MainHeader.test.ts
@@ -26,10 +26,8 @@ describe("MainHeader", () => {
     const wrapper = shallowMount(MainHeader, {
       props: {
         menuItems,
-        lang: {
-          locale: "de_DE",
-          id: "de",
-        },
+        lang: "de",
+        translationsRoutes: { de: "/de", en: "/en" },
       },
     });